+++ changelog.md
@@ -6,6 +6,8 @@ Codion Change Log
 - Value.Builder.initialValue() renamed value.
 ### is.codion.common.model
 - Unused slf4j dependency removed.
+### is.codion.swing.common.model
+- DefaultFilterTableModel, null item rejection improved.
 ### is.codion.swing.common.ui
 - AbstractComponentBuilder.setInitialValue() removed, no longer necessary now that the component value is always created, which is now used to set the initial value.
 - SpinnerNumberValue now non-null and validates the value according to the model minimum and maximum, if specified.